refactor norm balls into a single file, change `test_prox_keeps_feasible_point` to `test_prox_preserves_feasible_point`
